lagi dan lagi commit

+++ routes/api.php
@@ -17,9 +17,9 @@ Route::middleware(['auth:sanctum'])->group(function () {
     Route::get('/user', [AuthController::class, 'user']);
     Route::get('/konten',[KontenController::class, 'index']);
     Route::get('/konten/{id}',[KontenController::class, 'show']);
-    Route::post('/konten/create', [KontenController::class, 'store']);
-    Route::post('/konten/update/{id}', [KontenController::class, 'update']);
-    Route::delete('/konten/delete/{id}', [KontenController::class, 'destroy']);
+    Route::post('/konten/create', [KontenController::class, 'store'])->middleware('admin');
+    Route::post('/konten/update/{id}', [KontenController::class, 'update'])->middleware('admin');
+    Route::delete('/konten/delete/{id}', [KontenController::class, 'destroy'])->middleware('admin');
     Route::get('/konten/kategori/{kategori}', [KontenController::class, 'getKontenByKategori']);
     Route::get('/komen', [KomenController::class, 'index']);
     Route::get('/komen/{id}', [KomenController::class, 'show']);
